Update app.py
Completely rewrote the code. This version uses LangChain summarizer instead of using a version of GPT-3.5. We'll see if it performs as well. I also am using Streamlit for the interface as it's simpler for debugging the backend. I don't want to spend hours debugging the interface yet. I'm also going to add Whisper as the speech to text algorithm, but I need to get those details ironed out first.